plugins: properly destroy N sidechians
Now that we've tested this mostly successfully, this adds in the
remaining cleanup for the memory leaks left in for debugging. The last
remaining piece of missing functionality is a little bit mysterious:
secondary compressors don't apply when a source is paused or played, but
they do work when muted or unmuted. For our usecase this is fine, but
it's incomplete.