fix race condition in osx multithread monitor
The fix for this race condition introduces a big period of waiting for
the audio queue to clean itself up (10s) when removing audio monitoring
from a source. The audio playback stops immediately but the UI is
frozen.

I've included details for the issue below, but since we only /really/
care about the Linux version, I'm going to set this aside and move on to
fixing up resync and moving on to linux

The issue happens as a result of internal waiting on condition
variables. The actual blocking call is __psynch_cvwait but looking at
the disassembly in the callstack, it seems like the block is part of the
call to AQ::API::Queue::AwaitAllPendingCallbacks(bool). FWIW, that bool
seems to be set to false.

I'm gonna guess there's some condition variable / waitgroup to wait for
all pending callbacks to finish.

```
    0x1aaaeaeb0 <+500>: ldr    x0, [sp, #0x30]
    0x1aaaeaeb4 <+504>: cbz    w19, 0x1aaaeaecc          ; <+528>
    0x1aaaeaeb8 <+508>: mov    w1, #0x0
    0x1aaaeaebc <+512>: bl     0x1aaafa580               ; AQ::API::Queue::AwaitAllPendingCallbacks(bool)
```