invalidate reference to free'd memory
When the caller fails to establish connection, it will attempt to use
the reference established here to clean things up. This makes sense
during a shutdown operation of a real stream, but libsrt_open won't even
succeed in opening the connection. The logic down the line asserts the
reference is valid with a null check, this is not actually ever set.